requirements.txt
paho-mqtt==1.5.0
pkg-resources==0.0.0
smbus==1.1.post2
